All the basic files.
When I set up the SFDX project in the local github clone, it decided that EVERY class, LWC, or configuration file from the sandbox needed to be included in the repo. So I copied over just the ones used in this project to this repo because my account refuses to make a 9,000-file commit.